requirements_bi_unimed_goiania.txt
adal==1.2.7
alembic==1.7.4
amqp==5.0.6
anyio==3.3.3
apache-airflow==2.2.0
apache-airflow-providers-amazon==2.3.0
apache-airflow-providers-celery==2.1.0
apache-airflow-providers-cncf-kubernetes==2.0.3
apache-airflow-providers-docker==2.2.0
apache-airflow-providers-elasticsearch==2.0.3
apache-airflow-providers-ftp==2.0.1
apache-airflow-providers-google==6.0.0
apache-airflow-providers-grpc==2.0.1
apache-airflow-providers-hashicorp==2.1.1
apache-airflow-providers-http==2.0.1
apache-airflow-providers-imap==2.0.1
apache-airflow-providers-microsoft-azure==3.2.0
apache-airflow-providers-mysql==2.1.1
apache-airflow-providers-odbc==2.0.1
apache-airflow-providers-oracle==2.2.1
apache-airflow-providers-postgres==2.3.0
apache-airflow-providers-redis==2.0.1
apache-airflow-providers-sendgrid==2.0.1
apache-airflow-providers-sftp==2.1.1
apache-airflow-providers-slack==4.1.0
apache-airflow-providers-sqlite==2.0.1
apache-airflow-providers-ssh==2.2.0
apispec==3.3.2
argcomplete==1.12.3
attrs==20.3.0
Authlib==0.15.4
azure-batch==11.0.0
azure-common==1.1.27
azure-core==1.19.0
azure-cosmos==3.2.0
azure-datalake-store==0.0.52
azure-identity==1.6.1
azure-keyvault==4.1.0
azure-keyvault-certificates==4.3.0
azure-keyvault-keys==4.4.0
azure-keyvault-secrets==4.3.0
azure-kusto-data==0.0.45
azure-mgmt-containerinstance==1.5.0
azure-mgmt-core==1.3.0
azure-mgmt-datafactory==1.1.0
azure-mgmt-datalake-nspkg==3.0.1
azure-mgmt-datalake-store==0.5.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-resource==20.0.0
azure-nspkg==3.0.2
azure-storage-blob==12.8.1
azure-storage-common==2.1.0
azure-storage-file==2.1.0
Babel==2.9.1
backports.entry-points-selectable==1.1.0
bcrypt==3.2.0
billiard==3.6.4.0
blinker==1.4
boto3==1.18.58
botocore==1.21.58
cachetools==4.2.2
cattrs==1.5.0
celery==5.1.2
certifi==2020.12.5
cffi==1.14.6
charset-normalizer==2.0.6
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
clickclick==20.10.2
cloudpickle==1.4.1
colorama==0.4.4
colorlog==4.8.0
commonmark==0.9.1
croniter==1.0.15
cryptography==3.4.8
cx-Oracle==8.3.0
dask==2021.6.0
DateTime==4.4
decorator==5.1.0
defusedxml==0.7.1
dill==0.3.1.1
distlib==0.3.3
distributed==2.19.0
dnspython==2.1.0
docker==5.0.3
docutils==0.16
elasticsearch==7.13.4
elasticsearch-dbapi==0.2.6
elasticsearch-dsl==7.4.0
email-validator==1.1.3
et-xmlfile==1.1.0
eventlet==0.32.0
filelock==3.3.0
Flask==1.1.4
Flask-AppBuilder==3.3.3
Flask-Babel==1.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
flower==1.0.0
fsspec==2021.10.0
gevent==21.8.0
google-ads==14.0.0
google-api-core==1.31.3
google-api-python-client==1.12.8
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-appengine-logging==0.1.5
google-cloud-audit-log==0.1.1
google-cloud-automl==2.4.2
google-cloud-bigquery==2.28.1
google-cloud-bigquery-datatransfer==3.3.4
google-cloud-bigquery-storage==2.9.1
google-cloud-bigtable==1.7.0
google-cloud-build==3.5.2
google-cloud-container==1.0.1
google-cloud-core==1.7.2
google-cloud-datacatalog==3.4.3
google-cloud-dataproc==2.5.0
google-cloud-dlp==1.0.0
google-cloud-kms==2.8.0
google-cloud-language==1.3.0
google-cloud-logging==2.6.0
google-cloud-memcache==1.0.0
google-cloud-monitoring==2.6.0
google-cloud-os-login==2.3.2
google-cloud-pubsub==2.8.0
google-cloud-redis==2.2.4
google-cloud-secret-manager==1.0.0
google-cloud-spanner==1.19.1
google-cloud-speech==1.3.2
google-cloud-storage==1.42.3
google-cloud-tasks==2.6.0
google-cloud-texttospeech==1.0.1
google-cloud-translate==1.7.0
google-cloud-videointelligence==1.16.1
google-cloud-vision==1.0.0
google-cloud-workflows==1.2.3
google-crc32c==1.3.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
graphviz==0.17
greenlet==1.1.2
grpc-google-iam-v1==0.12.3
grpcio==1.41.0
grpcio-gcp==0.2.2
gunicorn==20.1.0
h11==0.12.0
HeapDict==1.0.1
httpcore==0.13.7
httplib2==0.19.1
httpx==0.19.0
humanize==3.12.0
hvac==0.11.2
idna==3.2
inflection==0.5.1
iso8601==0.1.16
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
json-merge-patch==0.2
jsonpath-ng==1.5.3
jsonschema==3.2.0
kombu==5.1.0
kubernetes==11.0.0
lazy-object-proxy==1.4.3
ldap3==2.9.1
libcst==0.3.21
locket==0.2.1
lockfile==0.12.2
Mako==1.1.5
Markdown==3.3.4
MarkupSafe==1.1.1
marshmallow==3.13.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.23.1
msal==1.15.0
msal-extensions==0.3.0
msgpack==1.0.2
msrest==0.6.21
msrestazure==0.6.4
mypy-extensions==0.4.3
mysql-connector-python==8.0.26
mysqlclient==2.0.3
nox==2020.12.31
numpy==1.20.3
oauthlib==3.1.1
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
openpyxl==3.0.9
packaging==21.0
pandas==1.3.3
pandas-gbq==0.14.1
pandasql==0.7.3
paramiko==2.8.0
partd==1.2.0
pendulum==2.1.2
platformdirs==2.4.0
ply==3.11
portalocker==1.7.1
prison==0.2.1
prometheus-client==0.11.0
prompt-toolkit==3.0.20
proto-plus==1.18.1
protobuf==3.17.3
psutil==5.8.0
psycopg2-binary==2.9.1
py==1.10.0
pyarrow==4.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydata-google-auth==1.2.0
Pygments==2.10.0
PyJWT==1.7.1
PyNaCl==1.4.0
pyodbc==4.0.32
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.18.0
pysftp==0.2.9
python-daemon==2.3.0
python-dateutil==2.8.2
python-http-client==3.3.3
python-ldap==3.3.1
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.3
pytzdata==2020.1
PyYAML==5.4.1
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
rfc3986==1.5.0
rich==10.12.0
rsa==4.7.2
s3transfer==0.5.0
sendgrid==6.8.2
setproctitle==1.2.2
six==1.16.0
slack-sdk==3.11.2
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.3.24
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.37.8
sshtunnel==0.1.5
starkbank-ecdsa==2.0.0
statsd==3.3.0
swagger-ui-bundle==0.0.9
tabulate==0.8.9
tblib==1.7.0
tenacity==8.0.1
termcolor==1.1.0
text-unidecode==1.3
timeout==0.1.2
toolz==0.11.1
tornado==6.1
typing-extensions==3.10.0.2
typing-inspect==0.7.1
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.26.7
vine==5.0.0
virtualenv==20.8.1
watchtower==1.0.6
wcwidth==0.2.5
websocket-client==1.2.1
Werkzeug==1.0.1
WTForms==2.3.3
zict==2.0.0
zope.event==4.5.0
zope.interface==5.4.0
